refactor(stella-store): extract inline test module to tests.rs
stella-store/src/lib.rs exceeded its file-size ratchet allowlist entry
(3970 lines, limit 3954) after recent additions, breaking main's CI gate.
The crate already splits large inline test modules into sibling files
(private_state_tests.rs, quarantine_tests.rs) — apply the same pattern to
the remaining ~1900-line inline `mod tests { ... }` block.

Pure mechanical relocation: the module body moves byte-for-byte (only
dedented, since it's no longer nested inside `mod tests { }`) into
stella-store/src/tests.rs, declared via `#[cfg(test)] mod tests;` in
lib.rs alongside its siblings. No behavior change — no witness test
needed for a pure refactor, per AGENTS.md.

Lowers both ratchet entries to lock in the gain: lib.rs 3904 -> 2035,
plus a new tests.rs entry at 1927 (itself over the 1500-line new-file cap,
same precedent as stella-pipeline/src/pipeline/tests.rs).

Part of the co-landing ratchet-union fix (see #270).
